adding license information for cdk code
